+++ fundamentals/today-i-learned/src/components/features/challenge/types.ts
@@ -1,0 +1,12 @@
+export interface ChallengeDay {
+  day: number;
+  date?: string;
+  status: "completed" | "today" | "pending" | "posted";
+  streak?: number;
+}
+
+export interface MonthlyChallenge {
+  year: number;
+  month: number;
+  days: ChallengeDay[];
+}
