+++ fundamentals/today-i-learned/src/components/features/profile/types.ts
@@ -1,0 +1,15 @@
+// === Streak & Activity Types ===
+
+export interface ActivityDay {
+  date: string;
+  hasActivity: boolean;
+  postCount: number;
+}
+
+export interface StreakData {
+  currentStreak: number;
+  longestStreak: number;
+  postsThisWeek: number;
+  totalPosts: number;
+  recentActivity: ActivityDay[];
+}
