+++ fundamentals/today-i-learned/src/components/shared/ui/ReactionTooltip.tsx
@@ -1,4 +1,5 @@
 import { ReactNode } from "react";
+import { css, cx } from "@styled-system/css";
 
 interface ReactionTooltipProps {
   isVisible: boolean;
@@ -6,23 +7,34 @@ interface ReactionTooltipProps {
   className?: string;
 }
 
+const tooltipStyles = css({
+  position: "absolute",
+  bottom: "full",
+  left: "50%",
+  transform: "translateX(-50%)",
+  marginBottom: "2",
+  display: "flex",
+  flexDirection: "row",
+  alignItems: "center",
+  backgroundColor: "white",
+  borderRadius: "full",
+  zIndex: 50,
+  padding: "1",
+  boxShadow: "0px 0px 10px rgba(0, 0, 0, 0.16)",
+  minWidth: "fit-content",
+  whiteSpace: "nowrap",
+  maxWidth: "300px"
+});
+
 export function ReactionTooltip({ 
   isVisible, 
   children,
-  className = "" 
+  className 
 }: ReactionTooltipProps) {
   if (!isVisible) return null;
 
   return (
-    <div
-      className={`absolute bottom-full left-1/2 transform -translate-x-1/2 mb-2 flex flex-row items-center bg-white rounded-full z-50 p-1 ${className}`}
-      style={{ 
-        boxShadow: "0px 0px 10px rgba(0, 0, 0, 0.16)",
-        minWidth: "fit-content",
-        whiteSpace: "nowrap",
-        maxWidth: "300px"
-      }}
-    >
+    <div className={cx(tooltipStyles, className)}>
       {children}
     </div>
   );
